+++ core/components/daterangetv/lexicon/de/default.inc.php
@@ -5,4 +5,11 @@
  * @package daterangetv
  * @subpackage lexicon
  */
-$_lang['daterange'] = 'Daterange';
+$_lang['daterange'] = 'Zeitraum';
+
+$_lang['setting_daterangetv.format'] = 'Datumsformat';
+$_lang['setting_daterangetv.format_desc'] = 'Eine zwischen Tag, Monat und Jahr durch | getrennte Liste von <a href="http://php.net/manual/en/function.strftime.php" target="_blank">strftime</a> Platzhaltern.';
+$_lang['setting_daterangetv.manager_format'] = 'Manager Date Format';
+$_lang['setting_daterangetv.manager_format_desc'] = 'Eine zwischen Tag, Monat und Jahr durch | getrennte Liste von <a href="http://php.net/manual/en/function.date.php" target="_blank">PHP date</a> Platzhaltern.';
+$_lang['setting_daterangetv.separator'] = 'Separator';
+$_lang['setting_daterangetv.separator_desc'] = 'Zeichenkette zwischen dem ersten und dem zweiten Teil des Zeitraums.';