+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 Changelog
 ================================================================================
 
+- 1.2
+    - Added column renderer (for MIGX/Collections etc.)
+    - Added default output system settings
 - 1.1
     - Added snippet/output filter
-
 - 1.0.1
     - Removed debug code
-
 - 1.0
     - Initial public release for MODX Revolution